# Improve encryption/decryption performance by caching private key