Right now encryption and decryption requires loading the private key from local storage every time.
With this PR, it caches the private key instead, and updates it when needed.

Improvement is around ~2x in performance.

Before:
```
----- EThree.encrypt & EThree.decrypt -----
[Alice] Encrypted and signed: 'MIIC5AIBADCCAkIGCSq[...]Ibf1iDLyt+x8='. Took: 225ms
[Bob] Decrypted and verified: Hello Bob!. Took: 124ms
[Bob] Encrypted and signed: 'MIIC5AIGCSqGSIb3DQE[...]LURHCMw=='. Took: 171ms
[Alice] Decrypted and verified: Hello Alice!. Took: 104ms
```

After:
```
----- EThree.encrypt & EThree.decrypt -----
[Alice] Encrypted and signed: 'MIIC5AIBADC[...]Y2SI92Biyi9TVvjLSv0='. Took: 103ms
[Bob] Decrypted and verified: Hello Bob!. Took: 57ms
[Bob] Encrypted and signed: 'MIIC5AIBA[...]rSlhTAhJ7/NA=='. Took: 85ms
[Alice] Decrypted and verified: Hello Alice!. Took: 59ms
```